refactor(dashboard): split goals.ts + tools.ts components
goals.ts (483 lines) split into goals/ subdirectory:
- goal-helpers.ts (123) — types, signals, derived data, helper functions
- goal-components.ts (126) — GoalRow, HorizonGroup, FilterBar, GoalsSummary
- mdal-components.ts (60) — LoopRow
- kanban-components.ts (87) — KanbanCard, TaskBacklog
- planning.ts (129) — Planning main component

tools.ts (463 lines) split into tools/ subdirectory:
- tool-state.ts (105) — signals, constants, loadTools, helpers
- tool-inventory-row.ts (39) — InventoryRow
- tool-summary-view.ts (69) — ToolSummaryView
- tool-full-inventory.ts (213) — FullInventoryView
- tools-main.ts (77) — Tools main component

Original files are now thin barrel re-exports. All resulting files
are under 350 lines. No import changes needed in consumers (work.ts,
control.ts, tab-refresh.ts) since barrel re-exports preserve the
public API.

Closes #1518
